+++ js/data.js
@@ -1,0 +1,39 @@
+const NAME_ARRAY = ['Артем', 'Михаил', 'Елена', 'Кира', 'Ольга', 'Александр', 'Игорь', 'Роман', 'Мия', 'Рамзан'];
+
+const PHOTO_DESCRIPTION = ['Городская улица с пешеходами в дождливый день.',
+  'Пейзаж гор с заснеженными вершинами и ясным небом.',
+  'Детский парк с улыбающимися детьми на качелях.',
+  'Осенний лес с золотистыми и красными листьями.',
+  'Крупный план цветка с каплями росы на лепестках.',
+  'Морской пляж с волнами и отдыхающими людьми.',
+  'Старинный замок на вершине холма на фоне заката.',
+  'Поля с подсолнухами под голубым небом.',
+  'Ночной городской пейзаж с огнями небоскребов.',
+  'Молодая пара,гуляющая по улице в весенний день.',
+  'Живописный вид на реку с отражением деревьев.',
+  'Экзотическая птица сидит на ветке с яркими перьями.',
+  'Пустыня с песчаными дюнами под жарким солнцем.',
+  'Заснеженная деревня с уютными домиками и дымом из труб.',
+  'Спортивный автомобиль на гоночной трассе в движении.',
+  'Портрет улыбающегося ребенка на фоне цветочного поля.',
+  'Мост через реку в туманное утро.',
+  'Фейерверк над городом в праздничный вечер.',
+  'Ярмарка с разнообразными товарами и людьми в движении.',
+  'Кафе на открытом воздухе с посетителями за столиками.',
+  'Водопад в глубоком лесу с бурным потоком воды.',
+  'Звездное небо с Млечным Путем над горами.',
+  'Сельская дорога среди зеленых лугов летом.',
+  'Семья на пикнике в парке в солнечный день.',
+  'Кафе на открытом воздухе с посетителями за столиками.'
+];
+
+const MESSAGE_ARRAY = ['Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'];
+
+const NUMBER_OF_PHOTO_OBJECTS = 25;
+
+export{NAME_ARRAY, PHOTO_DESCRIPTION, MESSAGE_ARRAY,NUMBER_OF_PHOTO_OBJECTS};
